+++ libs/rxjs/array/src/types/search.ts
@@ -1,26 +1,4 @@
 /**
- * Method used to modify the input value for comparison
- *
- * @typeParam T The type of the value to be modified for comparison
- * @typeParam K The type of the value returned by the function
- *
- * @param value The value to be modified for comparison
- *
- * @example
- * ```ts
- * const modify: InputModifierFn<T> = (value: string): string => value.toUpperCase()
- * ```
- *
- * @example
- * ```ts
- * const modify: InputModifierFn<T, K> = (value: string): number => parseInt(value, 10)
- * ```
- *
- * @returns Value of the modifier function
- */
-export type InputModifierFn<T = any, K = any> = (value: T) => K;
-
-/**
  * Method used to sort an Array of values
  *
  * @typeParam T The Type of the value to sort