+++ libs/rxjs/boolean/src/index.ts
@@ -9,3 +9,4 @@ export { filterTruthy } from './lib/filter-truthy';
 export { fromBoolean } from './lib/from-boolean';
 export { lastTruthy } from './lib/last-truthy';
 export { luhnCheck } from './lib/luhn-check';
+export { flip } from './lib/flip';