+++ libs/rxjs/boolean/src/lib/flip.ts
@@ -1,8 +1,22 @@
+/**
+ * @packageDocumentation
+ * @module boolean
+ */
 import { MonoTypeOperatorFunction, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
 /**
- * This operator takes a boolean value and flips to the opposite value
+ * The `flip` operator simply flips a boolean value
+ *
+ * @example
+ * ```ts
+ * fromBoolean(false)
+ *  .pipe(flip())
+ *  .subscribe() // true
+ * ```
+ *
+ * @returns Boolean value flipped from the input
+ * @category RxJS Boolean Modify
  */
 export function flip(): MonoTypeOperatorFunction<boolean> {
   return (source: Observable<boolean>) => source.pipe(map((value) => !value));