+++ libs/rxjs/number/src/index.ts
@@ -10,6 +10,7 @@ export { filterIsNotNaN } from './lib/filter-is-not-nan';
 export { filterIsSafeInteger } from './lib/filter-is-safe-integer';
 export { filterOutOfRange } from './lib/filter-out-of-range';
 export { fromNumber } from './lib/from-number';
+export { fromFibonacci } from './lib/from-fibonacci';
 export { inRange } from './lib/in-range';
 export { isFinite } from './lib/is-finite';
 export { isInteger } from './lib/is-integer';