+++ libs/rxjs/random/src/index.ts
@@ -1,4 +1,4 @@
 export { fromRandom } from './lib/from-random';
 export { fromRandomInt } from './lib/from-random-integer';
 export { fromRandomStr } from './lib/from-random-string';
-export { fromCryptoRandom } from './lib/from-crypto-random';
+export { fromRandomCrypto } from './lib/from-random-crypto';