+++ libs/rxjs/random/src/types/from-string.ts
@@ -1,0 +1,6 @@
+export interface FromRandomStringOpts {
+  caps: boolean;
+  lower: boolean;
+  number: boolean;
+  special: boolean;
+}